+++ CHANGELOG.md
@@ -10,6 +10,7 @@ pre: "<b>6. </b>"
 #### Added 
 
 - Missing JSON specification (and therefore, documentation) for `collectErrors`. See [#244](https://github.com/CardanoSolutions/ogmios/issues/244).
+- NixOS module for Ogmios server. See [nix/test-nixos-configuration.nix](nix/test-nixos-configuration.nix) for example usage.
 
 #### Changed